Create a test suite for tesseract_collision to be used by external packages
