+++ tesseract/tesseract_collision/include/tesseract_collision/test_suite/data/blender_monkey.bt
