+++ tesseract/tesseract_collision/include/tesseract_collision/test_suite/data/box_2m.bt
