+++ tesseract/tesseract_collision/include/tesseract_collision/test_suite/data/plane_4m.stl
