+++ tesseract/tesseract_collision/include/tesseract_collision/test_suite/data/sphere_p25m.ply
