+++ tesseract/tesseract_collision/test/CMakeLists.txt
@@ -11,13 +11,12 @@ find_package(tesseract_scene_graph REQUIRED)
 
 macro(add_gtest test_name test_file)
   add_executable(${test_name} ${test_file})
-  target_link_libraries(${test_name} PRIVATE GTest::GTest GTest::Main ${PROJECT_NAME}_bullet ${PROJECT_NAME}_fcl tesseract::tesseract_geometry tesseract::tesseract_scene_graph console_bridge ${BULLET_LIBRARIES} ${Boost_LIBRARIES} ${OCTOMAP_LIBRARIES} ${LIBFCL_LIBRARIES})
+  target_link_libraries(${test_name} PRIVATE GTest::GTest GTest::Main ${PROJECT_NAME}_test_suite ${PROJECT_NAME}_bullet ${PROJECT_NAME}_fcl tesseract::tesseract_geometry tesseract::tesseract_scene_graph console_bridge ${BULLET_LIBRARIES} ${Boost_LIBRARIES} ${OCTOMAP_LIBRARIES} ${LIBFCL_LIBRARIES})
   tesseract_target_compile_options(${test_name} PRIVATE)
   tesseract_clang_tidy(${test_name})
   tesseract_code_coverage(${test_name} ALL EXCLUDE ${COVERAGE_EXCLUDE})
-  target_compile_definitions(${test_name} PRIVATE DATA_DIR="${CMAKE_SOURCE_DIR}/test")
   tesseract_gtest_discover_tests(${test_name})
-  add_dependencies(${test_name} ${PROJECT_NAME}_bullet ${PROJECT_NAME}_fcl)
+  add_dependencies(${test_name} ${PROJECT_NAME}_core ${PROJECT_NAME}_test_suite ${PROJECT_NAME}_bullet ${PROJECT_NAME}_fcl)
   add_dependencies(run_tests ${test_name})
 endmacro()
