+++ tesseract/tesseract_collision/test/collision_box_capsule_unit.cpp
@@ -3,156 +3,29 @@ TESSERACT_COMMON_IGNORE_WARNINGS_PUSH
 #include <gtest/gtest.h>
 TESSERACT_COMMON_IGNORE_WARNINGS_POP
 
-#include "tesseract_collision/bullet/bullet_discrete_simple_manager.h"
-#include "tesseract_collision/bullet/bullet_discrete_bvh_manager.h"
-#include "tesseract_collision/fcl/fcl_discrete_managers.h"
+#include <tesseract_collision/test_suite/collision_box_capsule_unit.hpp>
+#include <tesseract_collision/bullet/bullet_discrete_simple_manager.h>
+#include <tesseract_collision/bullet/bullet_discrete_bvh_manager.h>
+#include <tesseract_collision/fcl/fcl_discrete_managers.h>
 
 using namespace tesseract_collision;
-using namespace tesseract_geometry;
-
-void addCollisionObjects(DiscreteContactManager& checker)
-{
-  //////////////////////
-  // Add box to checker
-  //////////////////////
-  CollisionShapePtr box(new Box(1, 1, 1));
-  Eigen::Isometry3d box_pose;
-  box_pose.setIdentity();
-
-  CollisionShapesConst obj1_shapes;
-  tesseract_common::VectorIsometry3d obj1_poses;
-  obj1_shapes.push_back(box);
-  obj1_poses.push_back(box_pose);
-
-  checker.addCollisionObject("box_link", 0, obj1_shapes, obj1_poses);
-
-  /////////////////////////////////////////////
-  // Add thin box to checker which is disabled
-  /////////////////////////////////////////////
-  CollisionShapePtr thin_box(new Box(0.1, 1, 1));
-  Eigen::Isometry3d thin_box_pose;
-  thin_box_pose.setIdentity();
-
-  CollisionShapesConst obj2_shapes;
-  tesseract_common::VectorIsometry3d obj2_poses;
-  obj2_shapes.push_back(thin_box);
-  obj2_poses.push_back(thin_box_pose);
-
-  checker.addCollisionObject("thin_box_link", 0, obj2_shapes, obj2_poses, false);
-
-  /////////////////////////////////////////////////////////////////
-  // Add capsule to checker.
-  /////////////////////////////////////////////////////////////////
-  CollisionShapePtr capsule(new Capsule(0.25, 0.25));
-  Eigen::Isometry3d capsule_pose;
-  capsule_pose.setIdentity();
-
-  CollisionShapesConst obj3_shapes;
-  tesseract_common::VectorIsometry3d obj3_poses;
-  obj3_shapes.push_back(capsule);
-  obj3_poses.push_back(capsule_pose);
-
-  checker.addCollisionObject("capsule_link", 0, obj3_shapes, obj3_poses);
-}
-
-void runTest(DiscreteContactManager& checker)
-{
-  //////////////////////////////////////
-  // Test when object is in collision
-  //////////////////////////////////////
-  checker.setActiveCollisionObjects({ "box_link", "capsule_link" });
-  checker.setContactDistanceThreshold(0.1);
-
-  // Set the collision object transforms
-  tesseract_common::TransformMap location;
-  location["box_link"] = Eigen::Isometry3d::Identity();
-  location["capsule_link"] = Eigen::Isometry3d::Identity();
-  location["capsule_link"].translation()(0) = 0.2;
-  checker.setCollisionObjectsTransform(location);
-
-  // Perform collision check
-  ContactResultMap result;
-  checker.contactTest(result, ContactTestType::CLOSEST);
-
-  ContactResultVector result_vector;
-  flattenResults(std::move(result), result_vector);
-
-  EXPECT_TRUE(!result_vector.empty());
-  EXPECT_NEAR(result_vector[0].distance, -0.55, 0.0001);
-  EXPECT_NEAR(result_vector[0].nearest_points[0][2], result_vector[0].nearest_points[1][2], 0.001);
-
-  std::vector<int> idx = { 0, 1, 1 };
-  if (result_vector[0].link_names[0] != "box_link")
-    idx = { 1, 0, -1 };
-
-  EXPECT_NEAR(result_vector[0].nearest_points[idx[0]][0], 0.5, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points[idx[0]][1], 0.0, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points[idx[1]][0], -0.05, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points[idx[1]][1], 0.0, 0.001);
-  EXPECT_NEAR(result_vector[0].normal[0], idx[2] * 1.0, 0.001);
-  EXPECT_NEAR(result_vector[0].normal[1], idx[2] * 0.0, 0.001);
-  EXPECT_NEAR(result_vector[0].normal[2], idx[2] * 0.0, 0.001);
-
-  ////////////////////////////////////////////////
-  // Test object is out side the contact distance
-  ////////////////////////////////////////////////
-  location["capsule_link"].translation() = Eigen::Vector3d(0, 0, 1);
-  result = ContactResultMap();
-  result.clear();
-  result_vector.clear();
-  checker.setCollisionObjectsTransform(location);
-
-  checker.contactTest(result, ContactTestType::CLOSEST);
-  flattenResults(std::move(result), result_vector);
-
-  EXPECT_TRUE(result_vector.empty());
-
-  /////////////////////////////////////////////
-  // Test object inside the contact distance
-  /////////////////////////////////////////////
-  result = ContactResultMap();
-  result.clear();
-  result_vector.clear();
-
-  checker.setContactDistanceThreshold(0.251);
-  checker.contactTest(result, ContactTestType::CLOSEST);
-  flattenResults(std::move(result), result_vector);
-
-  EXPECT_TRUE(!result_vector.empty());
-  EXPECT_NEAR(result_vector[0].distance, 0.125, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points[0][0], result_vector[0].nearest_points[1][0], 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points[0][1], result_vector[0].nearest_points[1][1], 0.001);
-
-  idx = { 0, 1, 1 };
-  if (result_vector[0].link_names[0] != "box_link")
-    idx = { 1, 0, -1 };
-
-  EXPECT_NEAR(result_vector[0].nearest_points[idx[0]][2], 0.5, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points[idx[1]][2], 0.625, 0.001);
-  EXPECT_NEAR(result_vector[0].normal[0], idx[2] * 0.0, 0.0011);  // FCL Required the bump in tolerance
-  EXPECT_NEAR(result_vector[0].normal[1], idx[2] * 0.0, 0.0011);  // FCL Required the bump in tolerance
-  EXPECT_NEAR(result_vector[0].normal[2], idx[2] * 1.0, 0.0011);  // FCL Required the bump in tolerance
-}
 
 TEST(TesseractCollisionUnit, BulletDiscreteSimpleCollisionBoxCapsuleUnit)  // NOLINT
 {
   tesseract_collision_bullet::BulletDiscreteSimpleManager checker;
-  addCollisionObjects(checker);
-  runTest(checker);
+  test_suite::runTest(checker);
 }
 
 TEST(TesseractCollisionUnit, BulletDiscreteBVHCollisionBoxCapsuleUnit)  // NOLINT
 {
   tesseract_collision_bullet::BulletDiscreteBVHManager checker;
-  addCollisionObjects(checker);
-  runTest(checker);
+  test_suite::runTest(checker);
 }
 
 TEST(TesseractCollisionUnit, FCLDiscreteBVHCollisionBoxCapsuleUnit)  // NOLINT
 {
   tesseract_collision_fcl::FCLDiscreteBVHManager checker;
-  addCollisionObjects(checker);
-  runTest(checker);
+  test_suite::runTest(checker);
 }
 
 int main(int argc, char** argv)
